ChangeDoctorIDTypeFormIntToLong
交DoctorID的类型从int改为了long

+++ ClassViewModelToDomain/DrugDoctorDepartmentCost.cs
@@ -7,7 +7,7 @@ namespace ClassViewModelToDomain
 {
     public class DrugDoctorDepartmentCost
     {
-        public int DoctorID { get; set; }
+        public long DoctorID { get; set; }
         public string Doctor { get; set; }
         public int DepartmentID { get; set; }
         public string Department { get; set; }
